fix(frontend): resolve layout issues and improve functionality
- Fix drag-to-resize functionality by properly binding event handlers after DOM load
- Improve notification positioning to avoid UI conflicts (top: 70px, z-index: 9999)
- Update keyboard shortcuts to prevent conflicts (Ctrl+/ for help instead of ?)
- Add temporary AI panel toggle button and F2 shortcut for testing three-column layout
- Create modular layout-manager.js for better code organization
- Enhanced resize handle binding with proper error handling and logging
- Update shortcuts help to reflect new key bindings
