+++ src/index.html
@@ -57,6 +57,9 @@
             <button id="toggleModeBtn" class="control-btn" onclick="toggleViewMode()" aria-label="Toggle editor/preview">
               👁
             </button>
+            <button id="toggleAiBtn" class="control-btn" onclick="toggleAiPanel()" aria-label="Toggle AI panel (F2)" title="Toggle AI Panel">
+              🤖
+            </button>
           </div>
         </header>
         <div class="editor-content" id="editorContent">
@@ -116,7 +119,8 @@
           <div class="shortcut"><kbd>Ctrl</kbd> + <kbd>O</kbd> - Open vault</div>
           <div class="shortcut"><kbd>Ctrl</kbd> + <kbd>E</kbd> - Toggle editor/preview</div>
           <div class="shortcut"><kbd>F1</kbd> - Toggle file tree</div>
-          <div class="shortcut"><kbd>?</kbd> - Show/hide shortcuts</div>
+          <div class="shortcut"><kbd>F2</kbd> - Toggle AI panel</div>
+          <div class="shortcut"><kbd>Ctrl</kbd> + <kbd>/</kbd> - Show/hide shortcuts</div>
         </div>
         <button onclick="toggleShortcutsHelp()" class="btn-primary">Close</button>
       </div>
